Move expiration-related functions to separate file
